projects/02_rag_knowledge_assistant/requirements.txt
langchain
openai
chromadb
streamlit
python-dotenv
pypdf
tiktoken